requirements.txt
django==2.1.15
django-bananas==1.3.3
django-bitfield==2.0.0
django-common-helpers==0.9.0
django-cors-headers==2.4.0
django-cron==0.5.0
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.2.0
django-hosts==3.0
django-import-export==1.0.0
django-redis==4.11.0
django-redshift-backend==0.7
django-simple-history==2.1.0
django-waffle==0.14.0
djangorestframework==3.8.2
pytest-django==3.8.0
factory-boy==3.1.0